requirements.txt
streamlit>=1.35.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
groq>=0.9.0
openai>=1.30.0
requests>=2.31.0
pdfplumber>=0.11.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
plotly>=5.22.0
numpy>=1.26.0
torch>=2.1.0
transformers>=4.40.0